Peparse v2 patch, expose peparsing to third-parties
